[NFC] Refactor common logic for JS interface
Both Unsubtyping and GTO need to analyzer JS-called functions as well as imports and exports to analyze types that might flow in from or out to JS. Extract the common logic into a shared utility in a new js-utils.h header. Also move the related jsPossibleJSPrototypeField method to this header.
